chore: Renamed test files
